refactor: rename infrastructure to data and application to ui. removed components folder
