+++ src/app/libs/dashboards/dashboards.module.ts
@@ -1,5 +1,5 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
-import { APPLICATION_PROVIDERS, INFRASTRUCTURE_PROVIDERS } from './infrastructure';
+import { UI_PROVIDERS, DATA_PROVIDERS } from './data';
 import { DashboardsUiModule } from './ui';
 
 @NgModule({
@@ -10,8 +10,8 @@ export class DashboardsModule {
     return {
       ngModule: DashboardsModule,
       providers: [
-        ...INFRASTRUCTURE_PROVIDERS,
-        ...APPLICATION_PROVIDERS,
+        ...DATA_PROVIDERS,
+        ...UI_PROVIDERS,
       ],
     };
   }
