+++ src/app/libs/dashboards/data/tokens.ts
@@ -2,7 +2,7 @@ import { SettingsApi } from '../../settings/ui';
 import { DashboardsDataPort, DashboardsUiPort, DashboardsUseCase } from '../use-cases';
 import { DashboardsStoreService } from './store/dashboards-store.service';
 
-export const INFRASTRUCTURE_PROVIDERS = [
+export const DATA_PROVIDERS = [
   {
     provide: DashboardsDataPort,
     useExisting: DashboardsStoreService,
@@ -10,7 +10,7 @@ export const INFRASTRUCTURE_PROVIDERS = [
   }
 ];
 
-export const APPLICATION_PROVIDERS = [
+export const UI_PROVIDERS = [
   {
     provide: DashboardsUiPort,
     useExisting: DashboardsUseCase,
