+++ src/app/libs/dashboards/ui/dashboards-ui.module.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { DashboardsComponent } from './components/dashboards.component';
+import { DashboardsComponent } from './dashboards/dashboards.component';
 
 @NgModule({
   declarations: [DashboardsComponent],
