+++ src/app/libs/sidebar/use-cases/index.ts
@@ -6,7 +6,7 @@ export * from './ports/ui/sidebar-ui.port';
 /**
  * Infrastructure Ports
  */
-export * from './ports/infrastructure/sidebar-data.port';
+export * from './ports/data/sidebar-data.port';
 
 /**
  * Use Case implementations
