+++ src/app/libs/sidebar/use-cases/sidebar.use-case.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { SidebarDataPort } from './ports/infrastructure/sidebar-data.port';
+import { SidebarDataPort } from './ports/data/sidebar-data.port';
 import { SidebarUiPort } from './ports/ui/sidebar-ui.port';
 
 @Injectable({providedIn: 'root'})
