+++ src/app/libs/spaces/use-cases/index.ts
@@ -5,7 +5,7 @@ export * from './ports/ui/spaces-ui.port';
 /**
  * Infrastructure Ports
  */
-export * from './ports/infrastructure/spaces-data.port';
+export * from './ports/data/spaces-data.port';
 
 /**
  * Use Case implementations
